Refactor QR/QI functions.
